Log message values may satisfy error-object? or condition?
My rationale is that logging often shows up in the context of
exception handlers, and the cause of the exceptions should be logged.
There are going to be more producers of such messages than consumers,
so it is better to put the cost of handling conditions on the
consumers. This is also no longer a feature that would be expected to
fall on the SRFI-215 implementer, which was previously a potential
roadblock to being able to log conditions.